Move `load_test_environment` function to `test_tools.py`
