+++ packages/pangea-sdk/tests/test_tools.py
@@ -1,6 +1,26 @@
+import os
 import unittest
 
+from pangea.exceptions import PangeaException
+from pangea.tools import TestEnvironment
 from pangea.utils import canonicalize, canonicalize_nested_json, hash_ntlm
+
+
+def load_test_environment(service_name: str, default: TestEnvironment = TestEnvironment.LIVE) -> TestEnvironment:
+    service_name = service_name.replace("-", "_").upper()
+    var_name = f"SERVICE_{service_name}_ENV"
+    value = os.getenv(var_name)
+    if not value:
+        print(f"{var_name} is not set. Return default test environment value: {default}")
+        return default
+    elif value == "DEV":
+        return TestEnvironment.DEVELOP
+    elif value == "STG":
+        return TestEnvironment.STAGING
+    elif value == "LVE":
+        return TestEnvironment.LIVE
+    else:
+        raise PangeaException(f"{var_name} not allowed value: {value}")
 
 
 class TestTools(unittest.TestCase):
